Add IdentityInjector
This adds an identity injector (that does not inject anything) and an
associated lack of injector info (NoInjectorInfo). This, while likely
never useful for actual injection use, is useful for testing.
